[Refactor] HomeView 자식 뷰 분리 및 일부 로직 하드코딩 해제
- MonthlyCalendarView와 WeeklyCalendarView를 별도의 파일로 분리
- MonthlyCalendarView의 높이가 해당 월의 주 개수에 따라 동적으로 조절되도록 수정
- 'getFestivalByDate' 호출에서 하드코딩된 연도 '2024'를 'selectedYear'로 변경하여 연도가 동적으로 전달되도록 수정
- 불필요한 주석을 제거하고, 복잡한 로직이나 의미가 불분명한 변수에 대한 주석 추가
